+++ CHANGELOG.md
@@ -8,6 +8,7 @@ v2.4.0
 * Change the format of the addAttribute method in order to support more kinds of JVM attributes.
 * Support generic type signatures.
 * Support named method parameters.
+* Support parameter annotations.
 
 v2.3.0 (2022-06-12)
 ------